feat(design): the catalogue says where it is, and titles stop running
Three things a phone was missing.

The catalogue opened on a row of chips and nothing else — no title, no
current genre, no sign of which page you were on; the bottom bar's icon
is a grid, which says nothing on its own. The header the desktop had was
computed inside its own branch. It moves out, and portrait gets it too.

An empty catalogue answered with one grey sentence in the top-left corner
of two thousand points of black, which reads as a page that failed rather
than a page that is waiting. It sits in the middle now.

And card titles scrolled forever on touch, because nothing is ever
focused there. In a grid that sets nine titles moving on nine different
cycles, and every glance catches three mid-word: "y Tale for Grown-u"
does not read as a title scrolling, it reads as a title broken. Ellipsis
is what every catalogue on a phone does. Marquee stays where there is a
focus to follow — one card at a time, on a device you cannot simply tap.

The scroll affordance built for the tab bar becomes a shared component,
since the filter bar overflowed in exactly the same way.
