+++ app/src/commonMain/kotlin/fr/moovie/tv/ui/details/DetailsTabs.kt
@@ -1,24 +1,8 @@
 package fr.moovie.tv.ui.details
 
 import androidx.compose.foundation.background
-import fr.moovie.tv.ui.theme.MOOVIE_TEXT_DIM
-import androidx.compose.ui.graphics.vector.ImageVector
-import androidx.compose.ui.graphics.graphicsLayer
-import androidx.compose.runtime.getValue
-import androidx.compose.material3.Icon
-import androidx.compose.material.icons.automirrored.filled.KeyboardArrowRight
-import androidx.compose.material.icons.automirrored.filled.KeyboardArrowLeft
-import androidx.compose.material.icons.Icons
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.BoxScope
-import androidx.compose.animation.core.animateFloatAsState
-import androidx.compose.foundation.horizontalScroll
-import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
@@ -27,10 +11,6 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.drawWithContent
-import androidx.compose.ui.geometry.Offset
-import androidx.compose.ui.geometry.Size
-import androidx.compose.ui.graphics.Brush
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.Dp
@@ -42,6 +22,7 @@ import fr.moovie.tv.resources.details_tab_similar
 import fr.moovie.tv.resources.details_tab_trailers
 import fr.moovie.tv.ui.adaptive.HeightClass
 import fr.moovie.tv.ui.adaptive.LocalHeightClass
+import fr.moovie.tv.ui.components.BarreDefilante
 import fr.moovie.tv.ui.components.MoovieButton
 import fr.moovie.tv.ui.theme.MOOVIE_ACCENT
 import org.jetbrains.compose.resources.StringResource
@@ -105,108 +86,20 @@ internal fun DetailsTabs(
      */
     premierFocus: Modifier = Modifier,
 ) {
-    // **La barre défile quand elle ne tient pas — et le dit.**
-    //
-    // Quatre onglets réclament près de six cents points ; un téléphone en
-    // portrait en offre quatre cents. Le dernier — « En savoir plus » —
-    // sortait donc simplement de l'écran : sur une série, la fiche n'avait
-    // plus de casting ni de fiche technique du tout. Un `Row` ne déborde pas,
-    // il tronque.
-    //
-    // Le rendre défilant ne suffisait pas. Le dernier onglet visible se
-    // terminait net avant le bord, exactement comme se termine une liste
-    // complète : rien ne distinguait « c'est tout » de « il y en a encore ».
-    // Un dégradé posé sur le bord vers lequel il reste à aller éteint le
-    // libellé au lieu de le couper, et cette coupure douce **est** le signe
-    // qu'on lit comme « continue par là ».
-    //
-    // Il ne paraît que du côté où il y a quelque chose : arrivé au bout, le
-    // bord redevient franc et la barre dit qu'elle est finie. Sur un
-    // téléviseur, où les quatre tiennent, aucun des deux ne s'allume.
-    //
-    // **Le dégradé seul ne suffisait pas non plus.** Ce qui dépasse au bord,
-    // ce n'est pas le libellé de l'onglet suivant mais le rembourrage de son
-    // bouton — quelques dizaines de points de vide. Le voile s'appliquait donc
-    // à du noir et n'éteignait rien du tout : le bord restait aussi franc
-    // qu'une barre complète. D'où le chevron, qui ne dépend pas de ce qui se
-    // trouve dessous pour se voir. Décoratif : il montre le chemin, on ne
-    // l'appuie pas — le geste, c'est de faire glisser la barre.
-    val defilement = rememberScrollState()
-    val versLaGauche = defilement.value > 0
-    val versLaDroite = defilement.value < defilement.maxValue
-    // Étroit à dessein. Ce qui dépasse d'un onglet suivant se compte en
-    // dizaines de points ; un voile plus large que ce dépassement l'efface
-    // entièrement, et l'on retrouve un bord franc — le dégradé s'appliquait
-    // alors à du noir. Il doit éteindre le libellé, pas le supprimer.
-    val voile = 20.dp
-    Box(
-        modifier = modifier.drawWithContent {
-            drawContent()
-            val largeur = voile.toPx()
-            if (versLaGauche) {
-                drawRect(
-                    brush = Brush.horizontalGradient(
-                        colors = listOf(MOOVIE_BG, Color.Transparent),
-                        startX = 0f,
-                        endX = largeur,
-                    ),
-                    size = Size(largeur, size.height),
-                )
-            }
-            if (versLaDroite) {
-                drawRect(
-                    brush = Brush.horizontalGradient(
-                        colors = listOf(Color.Transparent, MOOVIE_BG),
-                        startX = size.width - largeur,
-                        endX = size.width,
-                    ),
-                    topLeft = Offset(size.width - largeur, 0f),
-                    size = Size(largeur, size.height),
-                )
-            }
-        },
-    ) {
-        Row(
-            modifier = Modifier.fillMaxWidth().horizontalScroll(defilement),
-            horizontalArrangement = Arrangement.spacedBy(8.dp),
-            verticalAlignment = Alignment.CenterVertically,
-        ) {
-            Spacer(Modifier.width(marge))
-            onglets.forEachIndexed { index, onglet ->
-                OngletFiche(
-                    onglet = onglet,
-                    actif = onglet == actif,
-                    onClick = { onSelect(onglet) },
-                    modifier = if (index == 0) premierFocus else Modifier,
-                )
-            }
-            Spacer(Modifier.width(marge))
+    // La barre déborde sur un téléphone : quatre onglets réclament près de six
+    // cents points, un portrait en offre quatre cents. Le dernier — « En savoir
+    // plus » — tombait hors de l'écran, emportant le casting et la fiche
+    // technique de toutes les séries. Comment elle le signale : voir
+    // [BarreDefilante].
+    BarreDefilante(modifier = modifier, marge = marge) {
+        onglets.forEachIndexed { index, onglet ->
+            OngletFiche(
+                onglet = onglet,
+                actif = onglet == actif,
+                onClick = { onSelect(onglet) },
+                modifier = if (index == 0) premierFocus else Modifier,
+            )
         }
-        Chevron(Icons.AutoMirrored.Filled.KeyboardArrowLeft, versLaGauche, Alignment.CenterStart)
-        Chevron(Icons.AutoMirrored.Filled.KeyboardArrowRight, versLaDroite, Alignment.CenterEnd)
-    }
-}
-
-/**
- * Le repère de défilement, à un bord de la barre d'onglets.
- *
- * Il apparaît et disparaît en fondu plutôt que d'un coup : la barre bouge sous
- * le doigt, et une icône qui s'allume sèchement pendant ce mouvement se lit
- * comme un élément de plus, pas comme un état de celui qu'on manipule.
- *
- * `MOOVIE_TEXT_DIM` et non la couleur des libellés : c'est un panneau, pas un
- * onglet, et deux blancs côte à côte les mettraient sur le même plan.
- */
-@Composable
-private fun BoxScope.Chevron(icone: ImageVector, visible: Boolean, cote: Alignment) {
-    val opacite by animateFloatAsState(if (visible) 1f else 0f, label = "chevronOnglets")
-    if (opacite > 0f) {
-        Icon(
-            imageVector = icone,
-            contentDescription = null,
-            tint = MOOVIE_TEXT_DIM,
-            modifier = Modifier.align(cote).size(20.dp).graphicsLayer { alpha = opacite },
-        )
     }
 }
